+++ R/aaa.R
@@ -12,3 +12,5 @@ check_named <- function(x, ..., allow_null = FALSE, arg = caller_arg(x), call = 
   }
   invisible(NULL)
 }
+
+http_methods <- c('get', 'head', 'post', 'put', 'delete', 'connect', 'options', 'trace', 'patch')